refactor(cli): split main.rs into per-subcommand modules
Consolidation stage 5, item 7. `crates/cljrs/src/main.rs` was 1,920 lines
holding ten subcommands, the environment bootstrap, the REPL, the test runner,
the async driver, and the native-library loader, with a 223-line `run_command`
whose `Compile` arm alone ran to 120 lines.

The split follows what the code is for rather than where it happened to sit:

* `cli.rs` — what is true of `cljrs` as a whole: the global flags, the miette
  hook, the tracing subscriber, the large-stack worker thread, the JIT and
  stats policy. `run_command` is now a ten-line dispatcher with no logic.
* `commands/{run,repl,compile,eval,ir,test,deps,build_native,lsp,nrepl}.rs` —
  one module per subcommand, each owning its own clap arguments and its `run`.
* `session.rs` — what more than one subcommand needs: `setup_globals` (runtime,
  stdlib, `cljrs.edn` wiring, JIT attach), the source-path helpers, `eval_in` /
  `eval_form`, and error formatting.
* `native.rs` — loading native code into a running environment.

Each subcommand's arguments move from an inline `Commands` variant into a
`#[derive(clap::Args)]` struct in its own module, which is what lets the
argument definitions live next to the code that consumes them. **Every
`--help` page is byte-identical to the previous binary** — root and all
fifteen subcommands, diffed.

`cljrs` gains a library target. The binary is a one-line shim over
`cli::main`, and the library is what makes the CLI's internals reachable from
`crates/cljrs/tests` — which stage 5's next commit needs, since `cljrs-dylib`'s
end-to-end tests come with it. It is a tool, not an embedding API, and the
crate docs say so.

One helper is shared rather than duplicated: `file_to_namespace` was private to
the test runner but `compile`'s entry-namespace fallback needs it too, so it
moves to `session.rs` beside the other source-path helpers. The async driver's
thread-local stops being reachable from `main` — `session::with_async_driver`
scopes it instead.

No behaviour change: 1155 tests pass (unchanged), `eval`, `run samples/graph`,
`run samples/core_async`, `compile` + run of samples/life, and
`ir build --ns clojure.core` (151 functions, 2 unsupported — the same figure as
stages 1 through 4) all match.
