+++ crates/cljrs/src/main.rs
@@ -1,1899 +1,6 @@
-use std::io::{self, BufRead, Write};
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::time::Instant;
-
-use clap::{Parser, Subcommand, ValueEnum};
-use miette::IntoDiagnostic as _;
-use tracing_subscriber::filter::Targets;
-
-use cljrs_eval::{Env, EvalError, GlobalEnv, eval};
-use cljrs_gc::GcConfig;
-use cljrs_stdlib::{self as cljrs_stdlib};
-use cljrs_value::Value;
-
-mod commands;
-mod extensions;
-use commands::ir::IrCommands;
-
-/// Default thread stack size: 64 MiB.
-const DEFAULT_STACK_SIZE: usize = 64 * 1024 * 1024;
-
-/// clojurust — a Rust-hosted dialect of the Clojure programming language.
-#[derive(Parser)]
-#[command(name = "cljrs", version, about, long_about = None)]
-struct Cli {
-    /// Thread stack size in megabytes (default: 64).
-    /// Increase if you hit stack overflows with deeply recursive code.
-    #[arg(
-        long,
-        global = true,
-        value_name = "MB",
-        help = "Set thread stack size (default 64MB)"
-    )]
-    stack_size_mb: Option<usize>,
-
-    #[arg(long, global = true, help = "Enable debug logging")]
-    debug: bool,
-
-    #[arg(long, global = true, help = "Enable trace logging (implies --debug)")]
-    trace: bool,
-
-    /// Require valid GPG or SSH signatures on every versioned commit before
-    /// executing historical code.  Off by default.  Can also be enabled
-    /// per-project via `:verify-commit-signatures true` in `cljrs.edn`.
-    #[arg(long, global = true)]
-    verify_commit_signatures: bool,
-
-    /// Make pinned lookups of native (Rust-backed) functions an error when
-    /// the package's recorded provenance does not match the requested
-    /// commit (default: warn once per pin).  Can also be enabled
-    /// per-project via `:enforce-native-versions true` in `cljrs.edn`.
-    #[arg(long, global = true)]
-    enforce_native_versions: bool,
-
-    /// JIT compilation invocation threshold (default 1000; 0 = disable JIT).
-    /// Also read from CLJRS_JIT_THRESHOLD env var.
-    #[arg(
-        long,
-        global = true,
-        value_name = "N",
-        help = "JIT invocation threshold (0 to disable, default 1000)"
-    )]
-    jit_threshold: Option<u32>,
-
-    /// Warm threshold: tree-walked calls before a function is lowered to IR
-    /// in the background (default 50; 0 disables background lowering).
-    /// Also read from CLJRS_IR_THRESHOLD env var.  Applies even when the
-    /// JIT is disabled.
-    #[arg(
-        long,
-        global = true,
-        value_name = "N",
-        help = "Background IR lowering threshold (0 to disable, default 50)"
-    )]
-    ir_threshold: Option<u32>,
-
-    /// Feature-level logging flags: -X debug:gc,jit or -X trace:env
-    ///
-    /// Format: <level>:<feature1>,<feature2>,...
-    /// Levels: debug, trace.  Features: gc, env, ir, jit.
-    #[arg(short = 'X', global = true, value_name = "LEVEL:FEATURES")]
-    x_flags: Vec<String>,
-
-    /// Print GC statistics on exit. Pass a path to write them to a file;
-    /// pass the flag without a value to write them to stdout. Only the
-    /// `run`, `eval`, and `test` subcommands honour this flag.
-    #[arg(
-        long = "gc-stats",
-        global = true,
-        value_name = "FILE",
-        num_args = 0..=1,
-        default_missing_value = "",
-    )]
-    gc_stats: Option<String>,
-
-    /// Print JIT specialization / inline-cache statistics on exit (Phase
-    /// 10.6).  Pass a path to write them to a file; pass the flag without a
-    /// value to write them to stdout.  Only the `run`, `eval`, and `test`
-    /// subcommands honour this flag.
-    #[arg(
-        long = "jit-stats",
-        global = true,
-        value_name = "FILE",
-        num_args = 0..=1,
-        default_missing_value = "",
-    )]
-    jit_stats: Option<String>,
-
-    #[command(subcommand)]
-    command: Commands,
-}
-
-/// Code-generation target for `cljrs compile`.
-#[derive(Clone, Copy, PartialEq, Eq, Debug, ValueEnum)]
-enum CompileTarget {
-    /// Standalone binary via Cranelift.
-    Native,
-    /// WebAssembly module via the AOT wasm backend.
-    Wasm,
-}
-
-#[derive(Subcommand)]
-enum Commands {
-    /// Interpret a .cljrs or .cljc source file.
-    Run {
-        /// Path to the source file.
-        file: PathBuf,
-        /// Source directories to search when resolving `require`.
-        #[arg(long = "src-path", value_name = "DIR")]
-        src_paths: Vec<PathBuf>,
-        /// GC soft memory limit in MB (triggers collection when exceeded).
-        #[arg(long)]
-        gc_soft_limit_mb: Option<usize>,
-        /// GC hard memory limit in MB (forces collection when exceeded).
-        #[arg(long)]
-        gc_hard_limit_mb: Option<usize>,
-        /// Arguments forwarded to -main (everything after `--`).
-        #[arg(last = true, value_name = "ARGS")]
-        args: Vec<String>,
-    },
-    /// Start an interactive REPL.
-    Repl {
-        /// Source directories to search when resolving `require`.
-        #[arg(long = "src-path", value_name = "DIR")]
-        src_paths: Vec<PathBuf>,
-        /// GC soft memory limit in MB (triggers collection when exceeded).
-        #[arg(long)]
-        gc_soft_limit_mb: Option<usize>,
-        /// GC hard memory limit in MB (forces collection when exceeded).
-        #[arg(long)]
-        gc_hard_limit_mb: Option<usize>,
-    },
-    /// AOT-compile a source file or project to a native binary.
-    ///
-    /// When `cljrs.edn` is present in the working directory the compiler reads
-    /// `:paths` (source directories), `:deps` (dependency source roots), and
-    /// `:main` (entry-point namespace) from it automatically.  The `-main`
-    /// function is discovered from `:paths` only (not from dependencies); an
-    /// error is raised if more than one `-main` is found and neither `--main`
-    /// nor `:main` resolves the ambiguity.
-    Compile {
-        /// Path to the source file.  Optional when `cljrs.edn` is present and
-        /// the entry-point namespace can be determined from `--main`, `:main`
-        /// in `cljrs.edn`, or auto-detection of a unique `-main` function.
-        file: Option<PathBuf>,
-        /// Output path (a native binary, or a `.wasm` module with `--target wasm`).
-        #[arg(short, long)]
-        out: PathBuf,
-        /// Code-generation target: `native` (default) produces a standalone
-        /// binary via Cranelift; `wasm` produces a WebAssembly module via the
-        /// AOT wasm backend (the entry namespace's functions; the `"rt"` imports
-        /// are satisfied by the runtime built for `wasm32-unknown-unknown`).
-        #[arg(long, value_enum, default_value_t = CompileTarget::Native, value_name = "TARGET")]
-        target: CompileTarget,
-        /// Source directories to search when resolving `require`.
-        #[arg(long = "src-path", value_name = "DIR")]
-        src_paths: Vec<PathBuf>,
-        /// Namespace containing the `-main` entry point (e.g. `my.app.core`).
-        /// Overrides `:main` in `cljrs.edn` and auto-detection.
-        #[arg(long = "main", value_name = "NS")]
-        main_ns: Option<String>,
-        /// Compile a test harness that runs all tests in the given file/directory.
-        #[arg(long)]
-        test: bool,
-        /// Fail the build if the binary would embed readable Clojure source
-        /// text (interpreted preambles, bundled namespaces).  Rejected with
-        /// `--test` and `--target wasm`, which do not run the audit.
-        #[arg(long = "require-fully-compiled")]
-        require_fully_compiled: bool,
-        /// GC soft memory limit in MB (triggers collection when exceeded).
-        #[arg(long)]
-        gc_soft_limit_mb: Option<usize>,
-        /// GC hard memory limit in MB (forces collection when exceeded).
-        #[arg(long)]
-        gc_hard_limit_mb: Option<usize>,
-    },
-    /// Evaluate a single Clojure expression and print the result.
-    Eval {
-        /// The expression to evaluate.
-        expr: String,
-    },
-    /// Inspect, pre-lower, and visualize clojurust's intermediate representation.
-    Ir {
-        #[command(subcommand)]
-        command: IrCommands,
-    },
-    /// Run clojure.test tests for one or more namespaces.
-    ///
-    /// If no namespaces are given, discovers and runs all test namespaces
-    /// found in the source paths.
-    Test {
-        /// Namespaces to test (e.g. my.app.core-test).
-        /// If omitted, all namespaces in --src-path are discovered.
-        namespaces: Vec<String>,
-        /// Source directories to search when resolving `require`.
-        #[arg(long = "src-path", value_name = "DIR")]
-        src_paths: Vec<PathBuf>,
-        /// Print each passing assertion (helps identify which test hangs).
-        #[arg(long, short)]
-        verbose: bool,
-        /// GC soft memory limit in MB (triggers collection when exceeded).
-        #[arg(long)]
-        gc_soft_limit_mb: Option<usize>,
-        /// GC hard memory limit in MB (forces collection when exceeded).
-        #[arg(long)]
-        gc_hard_limit_mb: Option<usize>,
-    },
-    /// Manage project dependencies declared in cljrs.edn.
-    ///
-    /// Git dependencies are cached in ~/.cljrs/cache/git/.
-    /// No network access occurs unless you run `cljrs deps fetch`.
-    Deps {
-        #[command(subcommand)]
-        command: DepsCommands,
-    },
-    /// Build the project's native Rust crate as a shared library.
-    ///
-    /// Reads the `:rust` key from `cljrs.edn`, runs `cargo build` in the
-    /// declared crate directory, and prints the path of the resulting
-    /// `.so` / `.dylib` / `.dll`.  The library is loaded automatically by
-    /// `cljrs run` and `cljrs repl` to register native functions before any
-    /// Clojure code is evaluated.
-    ///
-    /// The user crate must declare `crate-type = ["cdylib"]` (or
-    /// `["cdylib", "rlib"]`) and export a `#[no_mangle] pub extern "C" fn
-    /// cljrs_init(registry: *mut cljrs_interop::Registry)` entry point.
-    BuildNative {
-        /// Build in release mode instead of debug.
-        #[arg(long)]
-        release: bool,
-    },
-    /// Run the clojurust language server (LSP) over stdio.
-    ///
-    /// Speaks the Language Server Protocol on stdin/stdout for editor
-    /// integration, providing parse diagnostics and a document-symbol outline
-    /// for `.cljrs` / `.cljc` files. Editors normally launch this for you.
-    Lsp,
-    /// Start an nREPL server for editor integration (CIDER, Calva, Conjure).
-    ///
-    /// Listens for bencode-encoded nREPL messages over TCP and writes the
-    /// bound port to `.nrepl-port` in the current directory so clients can
-    /// auto-connect. Runs until interrupted.
-    Nrepl {
-        /// Port to listen on (0 = let the OS pick a free port).
-        #[arg(long, default_value_t = 0)]
-        port: u16,
-        /// Address to bind.
-        #[arg(long, default_value = "127.0.0.1")]
-        bind: String,
-        /// Source directories to search when resolving `require`.
-        #[arg(long = "src-path", value_name = "DIR")]
-        src_paths: Vec<PathBuf>,
-        /// GC soft memory limit in MB (triggers collection when exceeded).
-        #[arg(long)]
-        gc_soft_limit_mb: Option<usize>,
-        /// GC hard memory limit in MB (forces collection when exceeded).
-        #[arg(long)]
-        gc_hard_limit_mb: Option<usize>,
-    },
-}
-
-#[derive(Subcommand)]
-enum DepsCommands {
-    /// Clone or update git dependencies from cljrs.edn.
-    ///
-    /// Without a name, fetches every git dependency declared in the
-    /// nearest cljrs.edn.  With a name, fetches only that dependency.
-    Fetch {
-        /// Dependency name to fetch (fetches all if omitted).
-        name: Option<String>,
-    },
-    /// Show which dependencies are cached and which are missing.
-    Status,
-}
-
-/// Build GC config from CLI flags, or use defaults if not specified.
-fn build_gc_config(soft_limit_mb: Option<usize>, hard_limit_mb: Option<usize>) -> Arc<GcConfig> {
-    match (soft_limit_mb, hard_limit_mb) {
-        (Some(soft), Some(hard)) => Arc::new(GcConfig::with_limits(
-            soft * 1024 * 1024,
-            hard * 1024 * 1024,
-        )),
-        (Some(soft), None) => Arc::new(GcConfig::with_hard_limit(soft * 1024 * 1024)),
-        (None, Some(hard)) => Arc::new(GcConfig::with_hard_limit(hard * 1024 * 1024)),
-        (None, None) => Arc::new(GcConfig::new()),
-    }
-}
-
-/// Install the global tracing subscriber.
-///
-/// The default verbosity comes from `--debug`/`--trace`, with the noisy
-/// codegen crates pinned to `warn` and the runtime's own feature targets
-/// (`gc`, `env`, `ir`, `jit`) pinned off — see
-/// [`cljrs_runtime::logging::base_filter`].  `RUST_LOG` (in `tracing`
-/// target=level syntax, e.g. `RUST_LOG=info,cranelift_codegen=debug`) replaces
-/// those defaults entirely, so the IR dumps are still reachable when wanted.
-///
-/// `-X debug:gc,jit` is layered on last, so it survives a `RUST_LOG` that says
-/// nothing about those targets.
-fn init_tracing(cli: &Cli) -> miette::Result<()> {
-    let default_level = if cli.trace {
-        tracing::Level::TRACE
-    } else if cli.debug {
-        tracing::Level::DEBUG
-    } else {
-        tracing::Level::INFO
-    };
-
-    let mut filter = match std::env::var("RUST_LOG") {
-        Ok(spec) if !spec.trim().is_empty() => spec
-            .parse::<Targets>()
-            .unwrap_or_else(|_| cljrs_runtime::logging::base_filter(default_level)),
-        _ => cljrs_runtime::logging::base_filter(default_level),
-    };
-
-    for flag in &cli.x_flags {
-        filter = cljrs_runtime::logging::apply_x_flag(filter, flag)
-            .map_err(|e| miette::miette!("invalid -X flag: {e}"))?;
-    }
-
-    cljrs_runtime::logging::init(filter);
-    Ok(())
-}
+//! The `cljrs` binary. Everything it does lives in the library beside it, so
+//! the CLI's integration tests can reach the same code.
 
 fn main() -> miette::Result<()> {
-    miette::set_hook(Box::new(|_| {
-        Box::new(
-            miette::MietteHandlerOpts::new()
-                .terminal_links(true)
-                .build(),
-        )
-    }))
-    .into_diagnostic()?;
-
-    let cli = Cli::parse();
-    init_tracing(&cli)?;
-
-    let stack_size = cli
-        .stack_size_mb
-        .map(|mb| mb * 1024 * 1024)
-        .unwrap_or(DEFAULT_STACK_SIZE);
-
-    // Spawn the actual work on a thread with a larger stack to handle
-    // deeply recursive Clojure code (lazy-seq chains, recursive macros, etc.).
-    let builder = std::thread::Builder::new()
-        .name("cljrs-main".into())
-        .stack_size(stack_size);
-    let handle = builder
-        .spawn(move || {
-            #[cfg(feature = "async")]
-            {
-                // Build the single-threaded runtime + LocalSet that every async
-                // task (core.async producers, ^:async calls, cljrs-io readers)
-                // runs on, and stash it so `eval_in` can drive each top-level
-                // form on it. We deliberately do *not* wrap `run` in a single
-                // `block_on`/`run_until`: top-level evaluation is synchronous
-                // (it reads stdin, dispatches commands), and each form is driven
-                // individually via `LocalSet::block_on`, which would panic if
-                // nested inside an outer `block_on`.
-                let rt = tokio::runtime::Builder::new_current_thread()
-                    .enable_all()
-                    .build()
-                    .expect("failed to build Tokio runtime");
-                let local = tokio::task::LocalSet::new();
-                ASYNC_DRIVER.with(|d| *d.borrow_mut() = Some(AsyncDriver { rt, local }));
-                let result = run(cli);
-                ASYNC_DRIVER.with(|d| *d.borrow_mut() = None);
-                result
-            }
-            #[cfg(not(feature = "async"))]
-            run(cli)
-        })
-        .into_diagnostic()?;
-    let result: miette::Result<i32> = handle.join().unwrap_or_else(|e| {
-        eprintln!("cljrs: thread panicked: {e:?}");
-        std::process::exit(1);
-    });
-    match result {
-        Ok(0) => Ok(()),
-        Ok(code) => std::process::exit(code),
-        Err(e) => Err(e),
-    }
-}
-
-fn run(cli: Cli) -> miette::Result<i32> {
-    // Register the main thread as a GC mutator so the collector knows
-    // how many threads to wait for during stop-the-world collection.
-    let _mutator = cljrs_gc::register_mutator();
-
-    // Decide whether runtimes this process builds get a JIT tier.
-    configure_jit(cli.jit_threshold.as_ref().copied());
-
-    // Configure background IR lowering (Phase 10.7); independent of the JIT.
-    match cli.ir_threshold {
-        // 0 disables background lowering entirely (functions stay at
-        // tree-walk unless eagerly lowered or pre-built).
-        Some(0) => cljrs_eval::set_ir_threshold(u32::MAX),
-        Some(t) => cljrs_eval::set_ir_threshold(t),
-        None => {}
-    }
-
-    let gc_stats_target = cli.gc_stats.clone();
-    let jit_stats_target = cli.jit_stats.clone();
-    let versioning = VersioningFlags {
-        verify_commit_signatures: cli.verify_commit_signatures,
-        enforce_native_versions: cli.enforce_native_versions,
-    };
-    let supports_gc_stats = matches!(
-        &cli.command,
-        Commands::Run { .. } | Commands::Eval { .. } | Commands::Test { .. },
-    );
-
-    let result = run_command(cli.command, versioning);
-
-    if supports_gc_stats
-        && let Some(target) = gc_stats_target.as_deref()
-        && let Err(e) = write_gc_stats(target)
-    {
-        eprintln!("cljrs: failed to write GC stats: {e}");
-    }
-    if supports_gc_stats
-        && let Some(target) = jit_stats_target.as_deref()
-        && let Err(e) = write_jit_stats(target)
-    {
-        eprintln!("cljrs: failed to write JIT stats: {e}");
-    }
-
-    result
-}
-
-/// Write a snapshot of the JIT specialization / inline-cache counters
-/// (`cljrs_compiler::rt_abi::jit_stats`) to `target`.
-///
-/// An empty target (the flag was passed without a value) writes to stdout;
-/// any other value is treated as a filesystem path.
-fn write_jit_stats(target: &str) -> std::io::Result<()> {
-    let snapshot = cljrs_compiler::rt_abi::jit_stats::snapshot();
-    if target.is_empty() {
-        println!("{snapshot}");
-        Ok(())
-    } else {
-        std::fs::write(target, snapshot)
-    }
-}
-
-/// Whether runtimes built by this process get a JIT tier attached.
-static JIT_ENABLED: std::sync::atomic::AtomicBool = std::sync::atomic::AtomicBool::new(false);
-
-/// Decide the JIT policy from CLI flags and env vars.
-///
-/// JIT is enabled by default; disable with `CLJRS_NO_JIT=1` or `--jit-threshold 0`.
-/// The backend itself is attached per runtime, in [`setup_globals`].
-fn configure_jit(threshold: Option<u32>) {
-    if std::env::var("CLJRS_NO_JIT").is_ok() {
-        return;
-    }
-    if let Some(0) = threshold {
-        return; // Explicitly disabled via --jit-threshold 0.
-    }
-    if let Some(t) = threshold {
-        cljrs_eval::jit_state::set_jit_threshold(t);
-    }
-    JIT_ENABLED.store(true, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// CLI-level versioned-symbol policy flags, threaded into `setup_globals`.
-#[derive(Clone, Copy, Default)]
-struct VersioningFlags {
-    /// `--verify-commit-signatures` / `:verify-commit-signatures`.
-    verify_commit_signatures: bool,
-    /// `--enforce-native-versions` / `:enforce-native-versions`.
-    enforce_native_versions: bool,
-}
-
-fn run_command(command: Commands, versioning: VersioningFlags) -> miette::Result<i32> {
-    match command {
-        Commands::Run {
-            file,
-            src_paths,
-            gc_soft_limit_mb,
-            gc_hard_limit_mb,
-            args,
-        } => {
-            let src = std::fs::read_to_string(&file)
-                .map_err(|e| miette::miette!("{}: {}", file.display(), e))?;
-            let filename = file.display().to_string();
-            let gc_config = build_gc_config(gc_soft_limit_mb, gc_hard_limit_mb);
-            let globals = setup_globals(src_paths, gc_config, versioning);
-            run_source(&src, &filename, globals, &args)?;
-            Ok(0)
-        }
-        Commands::Repl {
-            src_paths,
-            gc_soft_limit_mb,
-            gc_hard_limit_mb,
-        } => {
-            let gc_config = build_gc_config(gc_soft_limit_mb, gc_hard_limit_mb);
-            let globals = setup_globals(src_paths, gc_config, versioning);
-            run_repl(globals);
-            Ok(0)
-        }
-        Commands::Compile {
-            file,
-            out,
-            target,
-            src_paths,
-            main_ns,
-            test,
-            require_fully_compiled,
-            gc_soft_limit_mb,
-            gc_hard_limit_mb,
-        } => {
-            let _gc_config = build_gc_config(gc_soft_limit_mb, gc_hard_limit_mb);
-
-            // Load cljrs.edn; silently absent is fine.
-            let deps_config = std::env::current_dir()
-                .ok()
-                .and_then(|cwd| cljrs_project::config::load_config(&cwd).ok().flatten());
-
-            let rust_config = deps_config.as_ref().and_then(|c| c.rust.clone());
-
-            // Build combined source paths: CLI flags first, then cljrs.edn :paths,
-            // then each dependency's source roots.
-            let mut all_src_paths = src_paths.clone();
-            if let Some(ref config) = deps_config {
-                for p in &config.paths {
-                    if !all_src_paths.contains(p) {
-                        all_src_paths.push(p.clone());
-                    }
-                }
-                for p in collect_dep_src_paths(config) {
-                    if !all_src_paths.contains(&p) {
-                        all_src_paths.push(p);
-                    }
-                }
-            }
-
-            if test && target == CompileTarget::Wasm {
-                return Err(miette::miette!(
-                    "--test is not supported with --target wasm yet"
-                ));
-            }
-            let opacity = resolve_opacity_policy(require_fully_compiled, test)
-                .map_err(|e| miette::miette!("{e}"))?;
-
-            if test {
-                let test_dir = file.as_deref().unwrap_or_else(|| std::path::Path::new("."));
-                cljrs_compiler::aot::compile_test_harness(test_dir, &out, &all_src_paths)
-                    .map_err(|e| miette::miette!("{e}"))?;
-            } else {
-                let entry_file = match file {
-                    Some(f) => f,
-                    None => {
-                        // Project mode: determine the entry namespace and locate its file.
-                        let project_paths: Vec<PathBuf> = deps_config
-                            .as_ref()
-                            .map(|c| c.paths.clone())
-                            .unwrap_or_default();
-
-                        let effective_ns = if let Some(ns) = main_ns {
-                            // --main CLI flag takes top priority.
-                            ns
-                        } else if let Some(ns) =
-                            deps_config.as_ref().and_then(|c| c.main_ns.as_deref())
-                        {
-                            // :main in cljrs.edn is second priority.
-                            ns.to_string()
-                        } else {
-                            // Auto-detect from project :paths only (not dep paths).
-                            if project_paths.is_empty() {
-                                return Err(miette::miette!(
-                                    "no source paths: add :paths to cljrs.edn or use --src-path"
-                                ));
-                            }
-                            let mains = find_main_namespaces(&project_paths);
-                            match mains.len() {
-                                0 => {
-                                    return Err(miette::miette!(
-                                        "no -main function found in source paths; \
-                                     specify --main or add :main to cljrs.edn"
-                                    ));
-                                }
-                                1 => mains.into_iter().next().unwrap().0,
-                                _ => {
-                                    let list = mains
-                                        .iter()
-                                        .map(|(ns, _)| ns.as_str())
-                                        .collect::<Vec<_>>()
-                                        .join(", ");
-                                    return Err(miette::miette!(
-                                        "multiple -main functions found ({list}); \
-                                         specify --main or add :main to cljrs.edn"
-                                    ));
-                                }
-                            }
-                        };
-
-                        ns_to_file(&effective_ns, &all_src_paths).ok_or_else(|| {
-                            miette::miette!(
-                                "could not find source file for namespace {effective_ns}"
-                            )
-                        })?
-                    }
-                };
-
-                // The compiler does not pick extensions; this build's feature
-                // set does, exactly as it does for an interpreted run.
-                let session = cljrs_compiler::extensions::CompileSession::new(
-                    all_src_paths.clone(),
-                    extensions::default_set(),
-                )
-                .rust_config(rust_config.clone())
-                .verify_commit_signatures(versioning.verify_commit_signatures)
-                .opacity(opacity);
-
-                if target == CompileTarget::Wasm {
-                    cljrs_compiler::aot::compile_file_to_wasm(&entry_file, &out, &session)
-                        .map_err(|e| miette::miette!("{e}"))?;
-                } else {
-                    cljrs_compiler::aot::compile_file(&entry_file, &out, &session)
-                        .map_err(|e| miette::miette!("{e}"))?;
-                }
-            }
-            Ok(0)
-        }
-        Commands::Eval { expr } => {
-            let gc_config = Arc::new(GcConfig::new());
-            let globals = setup_globals(Vec::new(), gc_config, versioning);
-            let result = eval_source(&expr, "<eval>", globals)?;
-            if result != Value::Nil {
-                println!("{}", result);
-            }
-            Ok(0)
-        }
-        Commands::Ir { command } => commands::ir::run(command),
-        Commands::Test {
-            namespaces,
-            src_paths,
-            verbose,
-            gc_soft_limit_mb,
-            gc_hard_limit_mb,
-        } => run_tests_command(
-            namespaces,
-            src_paths,
-            verbose,
-            gc_soft_limit_mb,
-            gc_hard_limit_mb,
-            versioning,
-        ),
-        Commands::Deps { command } => match command {
-            DepsCommands::Fetch { name } => run_deps_fetch(name),
-            DepsCommands::Status => run_deps_status(),
-        },
-        Commands::BuildNative { release } => run_build_native(release),
-        Commands::Lsp => {
-            // cljrs-lsp owns its own runtime; this is safe because `run` never
-            // enters the interpreter's async driver runtime.
-            cljrs_lsp::run_stdio_blocking()
-                .map_err(|e| miette::miette!("lsp server error: {e}"))?;
-            Ok(0)
-        }
-        Commands::Nrepl {
-            port,
-            bind,
-            src_paths,
-            gc_soft_limit_mb,
-            gc_hard_limit_mb,
-        } => {
-            let gc_config = build_gc_config(gc_soft_limit_mb, gc_hard_limit_mb);
-            let globals = setup_globals(src_paths, gc_config, versioning);
-            let addr: std::net::SocketAddr = format!("{bind}:{port}")
-                .parse()
-                .map_err(|e| miette::miette!("invalid bind address {bind}:{port}: {e}"))?;
-            let config = cljrs_nrepl::Config {
-                addr,
-                port_file: Some(PathBuf::from(".nrepl-port")),
-            };
-            let server = cljrs_nrepl::start(config, globals)?;
-            // Editors (CIDER, Calva) parse this exact line to auto-connect.
-            println!(
-                "nREPL server started on port {0} on host {bind} - nrepl://{bind}:{0}",
-                server.port()
-            );
-            // Evaluate forms through the CLI's async driver so core.async /
-            // ^:async code makes progress, exactly like `cljrs repl`.
-            server.serve_with(eval_form)?;
-            Ok(0)
-        }
-    }
-}
-
-/// Resolve the opacity policy a `compile` invocation runs under.
-///
-/// Both backends audit: `compile_file` for embedded source, and
-/// `compile_file_to_wasm` for units the module would omit.  `--test` does not:
-/// `compile_test_harness` bundles every test namespace as interpreted source
-/// unconditionally, so a strict policy there could never be satisfied and the
-/// combination is refused rather than left to fail confusingly.
-fn resolve_opacity_policy(
-    require_fully_compiled: bool,
-    test: bool,
-) -> Result<cljrs_compiler::aot::OpacityPolicy, String> {
-    if !require_fully_compiled {
-        return Ok(cljrs_compiler::aot::OpacityPolicy::Report);
-    }
-    if test {
-        return Err("--require-fully-compiled is not supported with --test: \
-                    the test harness is not audited for embedded source"
-            .to_string());
-    }
-    Ok(cljrs_compiler::aot::OpacityPolicy::RequireFullyCompiled)
-}
-
-/// Write a snapshot of `cljrs_gc::GC_STATS` to `target`.
-///
-/// An empty target (the flag was passed without a value) writes to stdout;
-/// any other value is treated as a filesystem path.
-fn write_gc_stats(target: &str) -> std::io::Result<()> {
-    let snapshot = cljrs_gc::GC_STATS.snapshot();
-    if target.is_empty() {
-        println!("{snapshot}");
-        Ok(())
-    } else {
-        std::fs::write(target, format!("{snapshot}\n"))
-    }
-}
-
-// ── Environment setup ─────────────────────────────────────────────────────────
-
-/// Create a fully initialised `GlobalEnv` with stdlib, user source paths, GC
-/// config, and any `cljrs.edn` found in the current working directory.
-///
-/// Paths declared in `:paths` of `cljrs.edn` are appended to `src_paths` (CLI
-/// flags take precedence).  The parsed `DepsConfig` is stored in
-/// `GlobalEnv.deps_config` so that versioned symbol resolution and the
-/// `deps fetch`/`deps status` commands share the same config object.
-fn setup_globals(
-    src_paths: Vec<PathBuf>,
-    gc_config: Arc<GcConfig>,
-    versioning: VersioningFlags,
-) -> Arc<GlobalEnv> {
-    let runtime = cljrs_runtime::Runtime::builder()
-        .execution_mode(cljrs_runtime::ExecutionMode::Tiered)
-        .source_paths(src_paths)
-        .gc_config(gc_config)
-        .build()
-        .unwrap_or_else(|e| {
-            eprintln!("failed to start the runtime: {e}");
-            std::process::exit(1);
-        });
-    // Attach the JIT tier to this runtime (unless disabled on the command
-    // line or by `CLJRS_NO_JIT`); nothing has been lowered yet, so no promotion
-    // can be missed.
-    if JIT_ENABLED.load(std::sync::atomic::Ordering::Relaxed) {
-        cljrs_compiler::jit::install(&runtime);
-    }
-    cljrs_stdlib::install(&runtime);
-    let globals = runtime.into_globals();
-    if versioning.verify_commit_signatures {
-        globals
-            .verify_commit_signatures
-            .store(true, std::sync::atomic::Ordering::Relaxed);
-    }
-    if versioning.enforce_native_versions {
-        globals.set_enforce_native_versions(true);
-    }
-    // Opt-in pinned native packages (:rust/load :dylib in cljrs.edn).
-    cljrs_dylib::install(&globals);
-    if let Ok(cwd) = std::env::current_dir() {
-        apply_deps_config(&globals, &cwd);
-    }
-    // Initialise async inside the LocalSet so `cljrs_async::init` can spawn its
-    // background GC-service task (it calls `spawn_local`, which requires a
-    // LocalSet context). The task persists on the LocalSet and is serviced by
-    // each subsequent per-form drive in `eval_form`.
-    #[cfg(feature = "async")]
-    ASYNC_DRIVER.with(|d| {
-        let guard = d.borrow();
-        let init = |g: &Arc<GlobalEnv>| {
-            cljrs_async::init(g);
-            cljrs_io::init(g);
-            #[cfg(feature = "net")]
-            cljrs_net::init(g);
-            #[cfg(feature = "charset")]
-            cljrs_charset::init(g);
-        };
-        match guard.as_ref() {
-            Some(drv) => drv.local.block_on(&drv.rt, async { init(&globals) }),
-            None => init(&globals),
-        }
-    });
-    #[cfg(feature = "base64")]
-    cljrs_base64::init(&globals);
-    globals
-}
-
-/// Load the nearest `cljrs.edn` and wire its data into `globals`.
-///
-/// Silently does nothing when no config file is found; prints a warning to
-/// stderr when the file exists but cannot be parsed.
-fn apply_deps_config(globals: &Arc<GlobalEnv>, cwd: &Path) {
-    match cljrs_project::config::load_config(cwd) {
-        Ok(Some(config)) => {
-            // Append edn :paths to the source-path list (CLI paths come first).
-            {
-                let mut paths = globals.source_paths.write().unwrap();
-                for p in &config.paths {
-                    if !paths.contains(p) {
-                        paths.push(p.clone());
-                    }
-                }
-            }
-            // Append each dependency's own source roots so a plain `require` of
-            // a dep's namespace resolves (git deps are materialized from the
-            // local cache — run `cljrs deps fetch` first; no network here).
-            add_dep_source_paths(globals, &config);
-            if config.verify_commit_signatures {
-                globals
-                    .verify_commit_signatures
-                    .store(true, std::sync::atomic::Ordering::Relaxed);
-            }
-            globals.load_trusted_signers(&config);
-            if config.enforce_native_versions {
-                globals.set_enforce_native_versions(true);
-            }
-            // Load the native shared library (if :rust is configured) so that
-            // native functions are registered before any Clojure code runs.
-            if let Some(rust_config) = &config.rust {
-                load_native_lib(rust_config, globals);
-            }
-            *globals.deps_config.write().unwrap() = Some(Arc::new(config));
-        }
-        Ok(None) => {}
-        Err(e) => eprintln!("cljrs: warning: could not load cljrs.edn: {e}"),
-    }
-}
-
-/// Append each declared dependency's own source roots to `globals.source_paths`
-/// so a plain `(require '[dep.ns :as …])` resolves namespaces provided by a
-/// dependency rather than only by the running binary or the local project.
-///
-/// - **Local deps** (`:local/root`) contribute paths from the directory on disk.
-/// - **Git deps** are materialized from the local bare cache at their pinned
-///   `:git/sha` (no network — `cljrs deps fetch` must have populated it).  A
-///   missing cache warns and is skipped rather than aborting the run.
-///
-/// Each dep's roots come from its own `cljrs.edn` `:paths` (resolved relative to
-/// the dep root), defaulting to `src/`.  Only directories that actually exist
-/// are added; pure-native deps (no Clojure source) contribute nothing here and
-/// are brought in by the native-`require` loader instead.
-fn add_dep_source_paths(globals: &Arc<GlobalEnv>, config: &cljrs_project::config::DepsConfig) {
-    for (name, dep) in &config.deps {
-        let root = match dep {
-            cljrs_project::config::Dependency::Local { root } => {
-                if root.is_dir() {
-                    root.clone()
-                } else {
-                    eprintln!(
-                        "cljrs: warning: local dep {name} not found at {}",
-                        root.display()
-                    );
-                    continue;
-                }
-            }
-            cljrs_project::config::Dependency::Git(git) => {
-                match cljrs_project::vcs::worktree_at_commit(&git.url, &git.sha) {
-                    Ok(p) => p,
-                    Err(e) => {
-                        eprintln!(
-                            "cljrs: warning: git dep {name} ({}) is not available ({e}); \
-                             run `cljrs deps fetch`",
-                            git.url
-                        );
-                        continue;
-                    }
-                }
-            }
-        };
-        let mut paths = globals.source_paths.write().unwrap();
-        for p in dep_source_paths(&root) {
-            if p.is_dir() && !paths.contains(&p) {
-                paths.push(p);
-            }
-        }
-    }
-}
-
-/// Collect source paths contributed by every declared dependency in `config`.
-///
-/// Git deps are resolved from the local cache (run `cljrs deps fetch` first);
-/// missing deps emit a warning and are skipped.
-fn collect_dep_src_paths(config: &cljrs_project::config::DepsConfig) -> Vec<PathBuf> {
-    let mut paths = Vec::new();
-    for (name, dep) in &config.deps {
-        let root = match dep {
-            cljrs_project::config::Dependency::Local { root } => {
-                if root.is_dir() {
-                    root.clone()
-                } else {
-                    eprintln!(
-                        "cljrs: warning: local dep {name} not found at {}",
-                        root.display()
-                    );
-                    continue;
-                }
-            }
-            cljrs_project::config::Dependency::Git(git) => {
-                match cljrs_project::vcs::worktree_at_commit(&git.url, &git.sha) {
-                    Ok(p) => p,
-                    Err(e) => {
-                        eprintln!(
-                            "cljrs: warning: git dep {name} ({}) is not available ({e}); \
-                             run `cljrs deps fetch`",
-                            git.url
-                        );
-                        continue;
-                    }
-                }
-            }
-        };
-        for p in dep_source_paths(&root) {
-            if p.is_dir() && !paths.contains(&p) {
-                paths.push(p);
-            }
-        }
-    }
-    paths
-}
-
-/// Scan `src_paths` for source files that define a top-level `-main` function.
-///
-/// Returns a list of `(namespace_name, file_path)` pairs, one per file that
-/// contains `(defn -main ...)` or `(defn- -main ...)` at the top level.
-/// Only looks at the reader-level form structure (no evaluation).
-fn find_main_namespaces(src_paths: &[PathBuf]) -> Vec<(String, PathBuf)> {
-    use cljrs_reader::form::FormKind;
-
-    let mut results = Vec::new();
-    for dir in src_paths {
-        if !dir.is_dir() {
-            continue;
-        }
-        let mut files = Vec::new();
-        collect_source_files(dir, &mut files);
-        for file in files {
-            let Ok(src) = std::fs::read_to_string(&file) else {
-                continue;
-            };
-            let mut parser = cljrs_reader::Parser::new(src, file.display().to_string());
-            let Ok(forms) = parser.parse_all() else {
-                continue;
-            };
-            let mut ns_name: Option<String> = None;
-            let mut has_main = false;
-            for form in &forms {
-                if let FormKind::List(parts) = &form.kind
-                    && let Some(head) = parts.first()
-                    && let FormKind::Symbol(s) = &head.kind
-                {
-                    if s == "ns" {
-                        if let Some(second) = parts.get(1)
-                            && let FormKind::Symbol(n) = &second.kind
-                        {
-                            ns_name = Some(n.clone());
-                        }
-                    } else if (s == "defn" || s == "defn-")
-                        && parts.len() >= 2
-                        && let FormKind::Symbol(n) = &parts[1].kind
-                        && n == "-main"
-                    {
-                        has_main = true;
-                    }
-                }
-            }
-            if has_main {
-                let ns = ns_name.unwrap_or_else(|| {
-                    // Fall back to file-based namespace name if no (ns ...) form.
-                    file_to_namespace(dir, &file).unwrap_or_default()
-                });
-                results.push((ns, file));
-            }
-        }
-    }
-    results
-}
-
-/// Recursively collect `.cljrs` and `.cljc` files under `dir`.
-fn collect_source_files(dir: &Path, out: &mut Vec<PathBuf>) {
-    let Ok(entries) = std::fs::read_dir(dir) else {
-        return;
-    };
-    let mut entries: Vec<_> = entries.filter_map(|e| e.ok()).collect();
-    entries.sort_by_key(|e| e.file_name());
-    for entry in entries {
-        let path = entry.path();
-        if path.is_dir() {
-            collect_source_files(&path, out);
-        } else if let Some(ext) = path.extension()
-            && (ext == "cljrs" || ext == "cljc")
-        {
-            out.push(path);
-        }
-    }
-}
-
-/// Translate a Clojure namespace name into a source file path by searching
-/// `src_paths` for `<ns/with/slashes>.cljrs` or `.cljc`.
-fn ns_to_file(ns: &str, src_paths: &[PathBuf]) -> Option<PathBuf> {
-    let rel: String = ns.replace('.', "/").replace('-', "_");
-    for dir in src_paths {
-        for ext in &["cljrs", "cljc"] {
-            let path = dir.join(format!("{rel}.{ext}"));
-            if path.exists() {
-                return Some(path);
-            }
-        }
-    }
-    None
-}
-
-/// The source roots a dependency at `root` contributes: its `cljrs.edn`
-/// `:paths` (already resolved to absolute paths relative to the dep root) if a
-/// parseable config is present, else the conventional `<root>/src`.
-fn dep_source_paths(root: &Path) -> Vec<PathBuf> {
-    let cfg_path = root.join("cljrs.edn");
-    if cfg_path.exists()
-        && let Ok(src) = std::fs::read_to_string(&cfg_path)
-        && let Ok(parsed) = cljrs_project::config::parse_config(&src, &cfg_path)
-        && !parsed.paths.is_empty()
-    {
-        return parsed.paths;
-    }
-    vec![root.join("src")]
-}
-
-// ── Source evaluation ─────────────────────────────────────────────────────────
-
-/// Evaluate all forms in `src`, printing nothing. Returns the last value.
-fn eval_source(src: &str, filename: &str, globals: Arc<GlobalEnv>) -> miette::Result<Value> {
-    let mut env = Env::new(globals, "user");
-    eval_in(&mut env, src, filename)
-}
-
-/// Run a source file: evaluate all top-level forms, then call `-main` if defined.
-fn run_source(
-    src: &str,
-    filename: &str,
-    globals: Arc<GlobalEnv>,
-    args: &[String],
-) -> miette::Result<()> {
-    let mut env = Env::new(globals, "user");
-    eval_in(&mut env, src, filename)?;
-    call_main_if_defined(&mut env, args)?;
-    Ok(())
-}
-
-/// Call `-main` in the current namespace if it is defined, passing `args` as
-/// individual string arguments. Silently skips if `-main` is not defined.
-fn call_main_if_defined(env: &mut Env, args: &[String]) -> miette::Result<()> {
-    // resolve returns nil for undefined symbols; swallow lookup errors defensively.
-    let resolved = eval_in(env, "(resolve '-main)", "<main-check>").unwrap_or(Value::Nil);
-    if resolved == Value::Nil {
-        return Ok(());
-    }
-    let escaped: Vec<String> = args.iter().map(|s| escape_clojure_string(s)).collect();
-    let call = format!("(-main {})", escaped.join(" "));
-    let result = eval_in(env, &call, "<main>")?;
-    // An `^:async` `-main` returns a `Future` immediately, with its body
-    // queued as a task on the shared `LocalSet`. Await that future so the body
-    // (and anything it spawns) runs to completion before the process exits;
-    // for a synchronous `-main` this is a no-op pass-through.
-    await_main_result(result)?;
-    Ok(())
-}
-
-/// Drive `value` to a settled value on the shared async `LocalSet`. If `value`
-/// is a `Future`/`Promise` (e.g. the result of an `^:async` `-main`), this
-/// yields until it resolves; any other value is returned unchanged.
-#[cfg(feature = "async")]
-fn await_main_result(value: Value) -> miette::Result<()> {
-    ASYNC_DRIVER.with(|d| {
-        let guard = d.borrow();
-        match guard.as_ref() {
-            Some(drv) => {
-                drv.local
-                    .block_on(&drv.rt, cljrs_async::eval_async::await_value(value))
-                    .map_err(format_eval_error)?;
-                Ok(())
-            }
-            // No driver installed: nothing to await against, so leave as-is.
-            None => Ok(()),
-        }
-    })
-}
-
-#[cfg(not(feature = "async"))]
-fn await_main_result(_value: Value) -> miette::Result<()> {
-    Ok(())
-}
-
-/// Produce a Clojure string literal (double-quoted, with escapes) for `s`.
-fn escape_clojure_string(s: &str) -> String {
-    let mut out = String::with_capacity(s.len() + 2);
-    out.push('"');
-    for ch in s.chars() {
-        match ch {
-            '"' => out.push_str("\\\""),
-            '\\' => out.push_str("\\\\"),
-            '\n' => out.push_str("\\n"),
-            '\r' => out.push_str("\\r"),
-            '\t' => out.push_str("\\t"),
-            c => out.push(c),
-        }
-    }
-    out.push('"');
-    out
-}
-
-/// Evaluate `src` in an existing `Env`. Returns the last value.
-fn eval_in(env: &mut Env, src: &str, filename: &str) -> miette::Result<Value> {
-    let mut parser = cljrs_reader::Parser::new(src.to_string(), filename.to_string());
-    let forms = parser.parse_all().map_err(miette::Report::from)?;
-
-    let mut result = Value::Nil;
-    for form in forms {
-        let _alloc_frame = cljrs_gc::push_alloc_frame();
-        result = eval_form(&form, env).map_err(format_eval_error)?;
-    }
-    Ok(result)
-}
-
-/// The Tokio runtime + `LocalSet` that drive async evaluation, set up once in
-/// `main` and reused for the lifetime of the process.
-#[cfg(feature = "async")]
-struct AsyncDriver {
-    rt: tokio::runtime::Runtime,
-    local: tokio::task::LocalSet,
-}
-
-#[cfg(feature = "async")]
-thread_local! {
-    static ASYNC_DRIVER: std::cell::RefCell<Option<AsyncDriver>> =
-        const { std::cell::RefCell::new(None) };
-}
-
-/// Evaluate a single top-level form.
-///
-/// With the `async` feature, the form is driven on the shared `LocalSet` via
-/// [`tokio::task::LocalSet::block_on`], so spawned tasks — core.async channel
-/// producers, `^:async` calls, the `cljrs-io` readers/writers — make progress
-/// and a top-level `await` resolves. Tasks that outlive a form (e.g. a producer
-/// feeding a channel `def`d in one REPL line and consumed in the next) stay
-/// queued on the `LocalSet` and continue on the next form's drive. Without the
-/// feature it is a plain synchronous `eval`.
-#[cfg(feature = "async")]
-#[allow(clippy::result_large_err)]
-fn eval_form(form: &cljrs_reader::Form, env: &mut Env) -> Result<Value, EvalError> {
-    ASYNC_DRIVER.with(|d| {
-        let guard = d.borrow();
-        match guard.as_ref() {
-            Some(drv) => drv
-                .local
-                .block_on(&drv.rt, cljrs_async::eval_async::eval_async(form, env)),
-            // No driver installed (shouldn't happen once `main` runs): fall back
-            // to a synchronous evaluation rather than panicking.
-            None => eval(form, env),
-        }
-    })
-}
-
-#[cfg(not(feature = "async"))]
-#[allow(clippy::result_large_err)]
-fn eval_form(form: &cljrs_reader::Form, env: &mut Env) -> Result<Value, EvalError> {
-    eval(form, env)
-}
-
-fn format_eval_error(e: EvalError) -> miette::Report {
-    match e {
-        EvalError::Thrown(val) => miette::miette!("Unhandled exception: {}", val),
-        EvalError::UnboundSymbol(s) => miette::miette!("Unable to resolve symbol: {}", s),
-        EvalError::Arity {
-            name,
-            expected,
-            got,
-        } => miette::miette!("Wrong number of args ({got}) passed to {name}; expected {expected}"),
-        EvalError::NotCallable(s) => miette::miette!("Not a function: {}", s),
-        EvalError::Runtime(msg) => miette::miette!("{}", msg),
-        EvalError::GasExhausted => miette::miette!("gas exhausted"),
-        EvalError::ForbiddenEffect(operation) => {
-            miette::miette!("effect forbidden in transaction function: {operation}")
-        }
-        EvalError::Read(e) => miette::Report::from(e),
-        EvalError::Recur(_) => miette::miette!("recur outside of loop/fn"),
-        EvalError::CommitSignatureVerificationFailed { commit, reason } => {
-            miette::miette!("commit {commit:?} failed signature verification: {reason}")
-        }
-    }
-}
-
-// ── Native library support ────────────────────────────────────────────────────
-
-/// Return the expected on-disk path for the shared library produced by
-/// `cargo build` inside `crate_dir`.
-///
-/// Respects cargo's workspace semantics: when `crate_dir` is a workspace
-/// member, cargo writes artifacts to `<workspace_root>/target/`, not
-/// `<crate_dir>/target/`. We ask cargo where its target directory is via
-/// `cargo metadata`. If that fails (no cargo on PATH, malformed manifest,
-/// etc.), we fall back to `<crate_dir>/target/` so the standalone-crate
-/// case still works.
-fn native_lib_path(crate_dir: &Path, crate_name: &str, release: bool) -> PathBuf {
-    let profile = if release { "release" } else { "debug" };
-    let lib_file = if cfg!(target_os = "windows") {
-        format!("{crate_name}.dll")
-    } else if cfg!(target_os = "macos") {
-        format!("lib{crate_name}.dylib")
-    } else {
-        format!("lib{crate_name}.so")
-    };
-    let target_dir = cargo_target_dir(crate_dir).unwrap_or_else(|| crate_dir.join("target"));
-    target_dir.join(profile).join(lib_file)
-}
-
-/// Ask `cargo metadata` for the target directory that cargo will actually use
-/// when building inside `crate_dir`. Returns `None` on any failure; the caller
-/// is expected to fall back to `<crate_dir>/target`.
-fn cargo_target_dir(crate_dir: &Path) -> Option<PathBuf> {
-    let output = std::process::Command::new("cargo")
-        .args([
-            "metadata",
-            "--format-version",
-            "1",
-            "--no-deps",
-            "--offline",
-        ])
-        .current_dir(crate_dir)
-        .output()
-        .ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    let stdout = std::str::from_utf8(&output.stdout).ok()?;
-    // Cargo's metadata JSON puts `target_directory` at the top level. We do
-    // a small targeted extract rather than pulling in a JSON dependency.
-    let key = r#""target_directory":""#;
-    let start = stdout.find(key)? + key.len();
-    let rest = &stdout[start..];
-    let mut end = None;
-    let bytes = rest.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        match bytes[i] {
-            b'\\' if i + 1 < bytes.len() => i += 2,
-            b'"' => {
-                end = Some(i);
-                break;
-            }
-            _ => i += 1,
-        }
-    }
-    let raw = &rest[..end?];
-    Some(PathBuf::from(json_unescape(raw)))
-}
-
-/// Decode the small subset of JSON string escapes that can appear in a
-/// `target_directory` path emitted by `cargo metadata`.
-fn json_unescape(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    let mut chars = s.chars();
-    while let Some(c) = chars.next() {
-        if c != '\\' {
-            out.push(c);
-            continue;
-        }
-        match chars.next() {
-            Some('\\') => out.push('\\'),
-            Some('"') => out.push('"'),
-            Some('/') => out.push('/'),
-            Some('n') => out.push('\n'),
-            Some('t') => out.push('\t'),
-            Some('r') => out.push('\r'),
-            Some(other) => {
-                out.push('\\');
-                out.push(other);
-            }
-            None => out.push('\\'),
-        }
-    }
-    out
-}
-
-/// Load the shared library declared by `rust_config` and call its `cljrs_init`
-/// entry point to register native functions into `globals`.
-///
-/// A missing library emits a warning and returns — callers of unregistered
-/// functions will get a runtime error rather than a startup crash, which is
-/// friendlier during development.
-fn load_native_lib(rust_config: &cljrs_project::config::RustConfig, globals: &Arc<GlobalEnv>) {
-    let Some(init_fn) = rust_config.init_fn.as_deref() else {
-        return;
-    };
-    let Some(crate_name) = rust_config.crate_name() else {
-        return;
-    };
-    // Symbol name is the last segment of the Rust path, e.g. "cljrs_init".
-    let sym_name = init_fn.rsplit("::").next().unwrap_or(init_fn);
-
-    let lib_path = native_lib_path(&rust_config.crate_dir, crate_name, false);
-    if !lib_path.exists() {
-        eprintln!(
-            "cljrs: native library not found at {} — run `cljrs build-native` first",
-            lib_path.display()
-        );
-        return;
-    }
-
-    // SAFETY: we own the process and are responsible for ensuring the library
-    // stays loaded (via mem::forget below) for the entire lifetime of globals.
-    unsafe {
-        let lib = match libloading::Library::new(&lib_path) {
-            Ok(l) => l,
-            Err(e) => {
-                eprintln!("cljrs: could not load {}: {e}", lib_path.display());
-                return;
-            }
-        };
-
-        // The exported symbol has C linkage and takes a raw pointer so it is
-        // callable across the FFI boundary without ABI assumptions.
-        let sym_bytes: Vec<u8> = format!("{sym_name}\0").into_bytes();
-        let init: libloading::Symbol<unsafe extern "C" fn(*mut cljrs_interop::Registry)> =
-            match lib.get(&sym_bytes) {
-                Ok(s) => s,
-                Err(e) => {
-                    eprintln!(
-                        "cljrs: could not find symbol {sym_name} in {}: {e}",
-                        lib_path.display()
-                    );
-                    return;
-                }
-            };
-
-        let mut registry = cljrs_interop::Registry::new(globals.clone());
-        init(&mut registry as *mut _);
-
-        // Prevent the library from being unloaded — its code must remain
-        // reachable as long as any registered NativeFn closures exist.
-        std::mem::forget(lib);
-    }
-    eprintln!(
-        "[build-native] loaded {} ({})",
-        lib_path.display(),
-        sym_name
-    );
-}
-
-/// Build the native Rust crate declared in `cljrs.edn` as a shared library.
-fn run_build_native(release: bool) -> miette::Result<i32> {
-    let cwd = std::env::current_dir().into_diagnostic()?;
-    let config = cljrs_project::config::load_config(&cwd)
-        .into_diagnostic()?
-        .ok_or_else(|| miette::miette!("no cljrs.edn found in or above the current directory"))?;
-
-    let rust_config = config
-        .rust
-        .as_ref()
-        .ok_or_else(|| miette::miette!("no :rust key found in cljrs.edn"))?;
-
-    let crate_name = rust_config
-        .crate_name()
-        .ok_or_else(|| miette::miette!(":rust has no :init function; cannot derive crate name"))?;
-
-    eprintln!(
-        "[build-native] building {} in {}",
-        crate_name,
-        rust_config.crate_dir.display()
-    );
-
-    let mut cmd = std::process::Command::new("cargo");
-    cmd.arg("build");
-    if release {
-        cmd.arg("--release");
-    }
-    cmd.current_dir(&rust_config.crate_dir);
-
-    let status = cmd.status().into_diagnostic()?;
-    if !status.success() {
-        return Err(miette::miette!("cargo build failed"));
-    }
-
-    let lib_path = native_lib_path(&rust_config.crate_dir, crate_name, release);
-    eprintln!("[build-native] built {}", lib_path.display());
-    println!("{}", lib_path.display());
-
-    Ok(0)
-}
-
-// ── Deps subcommand ───────────────────────────────────────────────────────────
-
-/// Fetch one or all git dependencies declared in the nearest `cljrs.edn`.
-fn run_deps_fetch(name: Option<String>) -> miette::Result<i32> {
-    let cwd = std::env::current_dir().into_diagnostic()?;
-    let config = cljrs_project::config::load_config(&cwd)
-        .into_diagnostic()?
-        .ok_or_else(|| miette::miette!("no cljrs.edn found in or above the current directory"))?;
-
-    if config.deps.is_empty() {
-        println!("No dependencies declared in cljrs.edn.");
-        return Ok(0);
-    }
-
-    // Collect (dep_name, dependency) pairs to process.
-    let to_fetch: Vec<(&str, &cljrs_project::config::Dependency)> = if let Some(ref n) = name {
-        match config.find_dep(n) {
-            Some(dep) => vec![(n.as_str(), dep)],
-            None => {
-                return Err(miette::miette!("dependency {:?} not found in cljrs.edn", n));
-            }
-        }
-    } else {
-        config.deps.iter().map(|(n, d)| (n.as_ref(), d)).collect()
-    };
-
-    let mut all_ok = true;
-    for (dep_name, dep) in to_fetch {
-        match dep {
-            cljrs_project::config::Dependency::Git(git_dep) => {
-                eprintln!("fetching {dep_name} ({})...", git_dep.url);
-                match cljrs_project::vcs::fetch_remote(&git_dep.url, &git_dep.sha) {
-                    Ok(path) => eprintln!("  ok → {}", path.display()),
-                    Err(e) => {
-                        eprintln!("  error: {e}");
-                        all_ok = false;
-                    }
-                }
-            }
-            cljrs_project::config::Dependency::Local { root } => {
-                if root.exists() {
-                    eprintln!("{dep_name}: local dep at {} — ok", root.display());
-                } else {
-                    eprintln!(
-                        "{dep_name}: local dep at {} — directory not found",
-                        root.display()
-                    );
-                    all_ok = false;
-                }
-            }
-        }
-    }
-
-    Ok(if all_ok { 0 } else { 1 })
-}
-
-/// Print the cache status of every dependency declared in the nearest `cljrs.edn`.
-fn run_deps_status() -> miette::Result<i32> {
-    let cwd = std::env::current_dir().into_diagnostic()?;
-    let config = cljrs_project::config::load_config(&cwd)
-        .into_diagnostic()?
-        .ok_or_else(|| miette::miette!("no cljrs.edn found in or above the current directory"))?;
-
-    if config.deps.is_empty() {
-        println!("No dependencies declared in cljrs.edn.");
-        return Ok(0);
-    }
-
-    let mut all_ok = true;
-    for (dep_name, dep) in &config.deps {
-        match dep {
-            cljrs_project::config::Dependency::Git(git_dep) => {
-                let cache_path = cljrs_project::vcs::cache_path_for_url(&git_dep.url);
-                let sha_present = cache_path.exists()
-                    && std::process::Command::new("git")
-                        .arg("-C")
-                        .arg(&cache_path)
-                        .arg("cat-file")
-                        .arg("-e")
-                        .arg(git_dep.sha.as_ref())
-                        .status()
-                        .map(|s| s.success())
-                        .unwrap_or(false);
-                if sha_present {
-                    println!(
-                        "{dep_name}: cached (sha: {}, url: {})",
-                        git_dep.sha, git_dep.url
-                    );
-                } else {
-                    println!(
-                        "{dep_name}: NOT cached — run `cljrs deps fetch` (sha: {}, url: {})",
-                        git_dep.sha, git_dep.url
-                    );
-                    all_ok = false;
-                }
-            }
-            cljrs_project::config::Dependency::Local { root } => {
-                if root.exists() {
-                    println!("{dep_name}: local dep at {} — ok", root.display());
-                } else {
-                    println!("{dep_name}: local dep at {} — NOT FOUND", root.display());
-                    all_ok = false;
-                }
-            }
-        }
-    }
-
-    Ok(if all_ok { 0 } else { 1 })
-}
-
-// ── Test runner ───────────────────────────────────────────────────────────────
-
-/// Result of running tests for a single namespace.
-struct NsTestResult {
-    ns: String,
-    pass: i64,
-    fail: i64,
-    error: i64,
-    test_count: i64,
-    /// None if tests ran; Some(msg) if the ns failed to load.
-    load_error: Option<String>,
-}
-
-fn run_tests_command(
-    namespaces: Vec<String>,
-    src_paths: Vec<PathBuf>,
-    verbose: bool,
-    gc_soft_limit_mb: Option<usize>,
-    gc_hard_limit_mb: Option<usize>,
-    versioning: VersioningFlags,
-) -> miette::Result<i32> {
-    let gc_config = build_gc_config(gc_soft_limit_mb, gc_hard_limit_mb);
-    let globals = setup_globals(src_paths, gc_config, versioning);
-
-    let namespaces = if namespaces.is_empty() {
-        // Read the final source paths (which may include cljrs.edn :paths).
-        let effective_paths = globals.source_paths.read().unwrap().clone();
-        let discovered = discover_namespaces(&effective_paths);
-        if discovered.is_empty() {
-            eprintln!("cljrs test: no test namespaces found in source paths");
-            return Ok(2);
-        }
-        eprintln!("Discovered {} test namespace(s).\n", discovered.len());
-        discovered
-    } else {
-        namespaces
-    };
-
-    let mut env = Env::new(globals, "user");
-
-    // Ensure clojure.test is loaded.
-    eval_in(&mut env, "(require 'clojure.test)", "<test>")?;
-
-    if verbose {
-        eval_in(
-            &mut env,
-            "(alter-var-root (var clojure.test/*verbose*) (constantly true))",
-            "<test>",
-        )?;
-    }
-
-    let start = Instant::now();
-    let mut results: Vec<NsTestResult> = Vec::new();
-
-    for ns in &namespaces {
-        let result = run_single_ns_tests(&mut env, ns);
-        // Remove the namespace after testing so its closures and form-trees can
-        // be reclaimed by GC.  Without this all 233 namespaces accumulate
-        // simultaneously and peak RSS can exceed 15 GB.
-        // Two force_collect calls are required: GC_INITIAL_LIVES=2 means an
-        // unreachable object survives one cycle in the grace period before being
-        // freed on the second cycle.
-        env.globals.namespaces.write().unwrap().remove(ns.as_str());
-        env.globals.loaded.lock().unwrap().remove(ns.as_str());
-        cljrs_eval::force_collect(&env);
-        cljrs_eval::force_collect(&env);
-        results.push(result);
-    }
-
-    let elapsed = start.elapsed();
-
-    // Print summary.
-    print_summary(&results, elapsed);
-
-    let total_fail: i64 = results.iter().map(|r| r.fail + r.error).sum();
-    let total_load_errors: usize = results.iter().filter(|r| r.load_error.is_some()).count();
-
-    if total_fail > 0 || total_load_errors > 0 {
-        Ok(1)
-    } else {
-        Ok(0)
-    }
-}
-
-fn run_single_ns_tests(env: &mut Env, ns: &str) -> NsTestResult {
-    // Try to load the namespace.
-    if let Err(e) = eval_in(env, &format!("(require '{ns})"), "<test>") {
-        return NsTestResult {
-            ns: ns.to_string(),
-            pass: 0,
-            fail: 0,
-            error: 0,
-            test_count: 0,
-            load_error: Some(format!("{e}")),
-        };
-    }
-
-    // Run the tests.
-    match eval_in(env, &format!("(clojure.test/run-tests '{ns})"), "<test>") {
-        Ok(counters) => {
-            let (pass, fail, error, test_count) = extract_counters(&counters);
-            NsTestResult {
-                ns: ns.to_string(),
-                pass,
-                fail,
-                error,
-                test_count,
-                load_error: None,
-            }
-        }
-        Err(e) => NsTestResult {
-            ns: ns.to_string(),
-            pass: 0,
-            fail: 0,
-            error: 0,
-            test_count: 0,
-            load_error: Some(format!("run-tests failed: {e}")),
-        },
-    }
-}
-
-fn extract_counters(val: &Value) -> (i64, i64, i64, i64) {
-    let Value::Map(m) = val else {
-        return (0, 0, 0, 0);
-    };
-    let mut pass = 0i64;
-    let mut fail = 0i64;
-    let mut error = 0i64;
-    let mut test_count = 0i64;
-    m.for_each(|k, v| {
-        if let (Value::Keyword(kw), Value::Long(count)) = (k, v) {
-            match kw.get().name.as_ref() {
-                "pass" => pass = *count,
-                "fail" => fail = *count,
-                "error" => error = *count,
-                "test" => test_count = *count,
-                _ => {}
-            }
-        }
-    });
-    (pass, fail, error, test_count)
-}
-
-fn print_summary(results: &[NsTestResult], elapsed: std::time::Duration) {
-    let total_tests: i64 = results.iter().map(|r| r.test_count).sum();
-    let total_assertions: i64 = results.iter().map(|r| r.pass + r.fail + r.error).sum();
-    let total_pass: i64 = results.iter().map(|r| r.pass).sum();
-    let total_fail: i64 = results.iter().map(|r| r.fail).sum();
-    let total_error: i64 = results.iter().map(|r| r.error).sum();
-    let load_errors: Vec<&NsTestResult> =
-        results.iter().filter(|r| r.load_error.is_some()).collect();
-    let ns_with_failures: Vec<&NsTestResult> = results
-        .iter()
-        .filter(|r| r.load_error.is_none() && (r.fail > 0 || r.error > 0))
-        .collect();
-
-    println!();
-    println!("══════════════════════════════════════════════════════════════");
-    println!("Test Summary");
-    println!("══════════════════════════════════════════════════════════════");
-    println!(
-        "Ran {} tests containing {} assertions across {} namespace(s) in {:.1}s.",
-        total_tests,
-        total_assertions,
-        results.len(),
-        elapsed.as_secs_f64()
-    );
-    println!(
-        "{} passed, {} failed, {} errors.",
-        total_pass, total_fail, total_error
-    );
-
-    if !load_errors.is_empty() {
-        println!();
-        println!(
-            "── {} namespace(s) failed to load ──────────────────────────────",
-            load_errors.len()
-        );
-        for r in &load_errors {
-            println!("  {} — {}", r.ns, r.load_error.as_deref().unwrap_or("?"));
-        }
-    }
-
-    if !ns_with_failures.is_empty() {
-        println!();
-        println!(
-            "── {} namespace(s) with test failures ──────────────────────────",
-            ns_with_failures.len()
-        );
-        for r in &ns_with_failures {
-            println!("  {} — {} failures, {} errors", r.ns, r.fail, r.error);
-        }
-    }
-
-    if load_errors.is_empty() && ns_with_failures.is_empty() {
-        println!();
-        println!("All tests passed.");
-    }
-    println!("══════════════════════════════════════════════════════════════");
-}
-
-/// Discover all namespace names from `.cljc` / `.cljrs` files in the given source paths.
-fn discover_namespaces(src_paths: &[PathBuf]) -> Vec<String> {
-    let mut namespaces = Vec::new();
-    for dir in src_paths {
-        if dir.is_dir() {
-            discover_in_dir(dir, dir, &mut namespaces);
-        }
-    }
-    namespaces.sort();
-    namespaces
-}
-
-fn discover_in_dir(root: &PathBuf, dir: &PathBuf, out: &mut Vec<String>) {
-    let Ok(entries) = std::fs::read_dir(dir) else {
-        return;
-    };
-    let mut entries: Vec<_> = entries.filter_map(|e| e.ok()).collect();
-    entries.sort_by_key(|e| e.file_name());
-    for entry in entries {
-        let path = entry.path();
-        if path.is_dir() {
-            discover_in_dir(root, &path, out);
-        } else if let Some(ext) = path.extension()
-            && (ext == "cljc" || ext == "cljrs")
-            && let Some(ns) = file_to_namespace(root, &path)
-        {
-            out.push(ns);
-        }
-    }
-}
-
-/// Convert a file path relative to the source root into a Clojure namespace name.
-/// e.g. `test/clojure/core_test/juxt.cljc` relative to `test/` → `clojure.core-test.juxt`
-fn file_to_namespace(root: &PathBuf, file: &Path) -> Option<String> {
-    let rel = file.strip_prefix(root).ok()?;
-    let stem = rel.with_extension(""); // remove .cljc / .cljrs
-    let ns = stem
-        .to_string_lossy()
-        .replace(std::path::MAIN_SEPARATOR, ".")
-        .replace('_', "-");
-    Some(ns)
-}
-
-// ── REPL ──────────────────────────────────────────────────────────────────────
-
-fn run_repl(globals: Arc<GlobalEnv>) {
-    println!("clojurust REPL (type :quit to exit)");
-    println!();
-
-    #[cfg(feature = "enable-rustyline")]
-    let mut rl = rustyline::DefaultEditor::new().unwrap();
-
-    let mut env = Env::new(globals, "user");
-
-    let stdin = io::stdin();
-    let mut input_buf = String::new();
-    let mut depth: i32 = 0;
-
-    #[cfg(feature = "enable-rustyline")]
-    loop {
-        let readline = rl.readline("=> ");
-        match readline {
-            Ok(line) => {
-                rl.add_history_entry(line.as_str());
-                if line.is_empty() {
-                    continue;
-                } else if line.starts_with(":quit") {
-                    break;
-                } else {
-                    match eval_in(&mut env, &line, "<repl>") {
-                        Ok(Value::Nil) => println!("nil"),
-                        Ok(v) => println!("{}", v),
-                        Err(e) => println!("error: {}", e),
-                    }
-                }
-            }
-            Err(rustyline::error::ReadlineError::Interrupted) => break,
-            Err(rustyline::error::ReadlineError::Eof) => break,
-            Err(err) => {
-                eprintln!("error: {}", err);
-                break;
-            }
-        }
-    }
-
-    #[cfg(not(feature = "enable-rustyline"))]
-    loop {
-        let prompt = if input_buf.is_empty() { "=> " } else { ".. " };
-        print!("{}", prompt);
-        io::stdout().flush().unwrap();
-
-        let mut line = String::new();
-        match stdin.lock().read_line(&mut line) {
-            Ok(0) => break, // EOF
-            Ok(_) => {}
-            Err(e) => {
-                eprintln!("I/O error: {e}");
-                break;
-            }
-        }
-
-        let trimmed = line.trim_end();
-
-        if input_buf.is_empty() && trimmed == ":quit" {
-            break;
-        }
-
-        // Track paren depth to support multi-line input.
-        for ch in trimmed.chars() {
-            match ch {
-                '(' | '[' | '{' => depth += 1,
-                ')' | ']' | '}' => depth -= 1,
-                _ => {}
-            }
-        }
-
-        if !input_buf.is_empty() {
-            input_buf.push('\n');
-        }
-        input_buf.push_str(trimmed);
-
-        // Only evaluate when parens are balanced (or we have a bare atom).
-        if depth <= 0 && !input_buf.trim().is_empty() {
-            depth = 0;
-            let src = std::mem::take(&mut input_buf);
-            match eval_in(&mut env, &src, "<repl>") {
-                Ok(Value::Nil) => {}
-                Ok(v) => println!("{}", v),
-                Err(e) => eprintln!("Error: {e}"),
-            }
-        }
-    }
-
-    println!("Bye.");
-}
-
-#[cfg(test)]
-mod tests {
-    use super::resolve_opacity_policy;
-    use cljrs_compiler::aot::OpacityPolicy;
-
-    /// Without the flag the policy is the reporting default.
-    #[test]
-    fn absent_flag_always_reports() {
-        for test in [false, true] {
-            assert_eq!(
-                resolve_opacity_policy(false, test),
-                Ok(OpacityPolicy::Report),
-                "test={test}"
-            );
-        }
-    }
-
-    #[test]
-    fn flag_selects_the_strict_policy_on_the_audited_path() {
-        assert_eq!(
-            resolve_opacity_policy(true, false),
-            Ok(OpacityPolicy::RequireFullyCompiled)
-        );
-    }
-
-    /// `--test` rejects rather than silently ignoring the flag.
-    #[test]
-    fn flag_is_rejected_under_test() {
-        let err = resolve_opacity_policy(true, true).expect_err("--test must be rejected");
-        assert!(
-            err.contains("--require-fully-compiled") && err.contains("--test"),
-            "error names both flags: {err}"
-        );
-    }
+    cljrs::cli::main()
 }
